+++ src/tests/testfiles/PostgreSqlAggregateMinimumTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceQuery.Xunit
+{
+    public class PostgreSqlAggregateMinimumTests : AggregateMinimumTests<TestClass>
+    {
+        public PostgreSqlAggregateMinimumTests()
+        {
+            ValidateUInt128 = false;
+        }
+
+        public override IQueryable<TestClass> GetTestList()
+        {
+            return PostgreSqlHelper.GetTestList();
+        }
+    }
+}