+++ src/tests/testfiles/PostgreSqlComparisonEndsWithTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceQuery.Xunit
+{
+    public class PostgreSqlComparisonEndsWithTests : ComparisonEndsWithTests<TestClass>
+    {
+        public PostgreSqlComparisonEndsWithTests()
+        {
+            ValidateUInt128 = false;
+        }
+
+        public override IQueryable<TestClass> GetTestList()
+        {
+            return PostgreSqlHelper.GetTestList();
+        }
+    }
+}