+++ src/tests/testfiles/PostgreSqlComparisonIsGreaterThanOrEqualTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceQuery.Xunit
+{
+    public class PostgreSqlComparisonIsGreaterThanOrEqualTests : ComparisonIsGreaterThanOrEqualTests<TestClass>
+    {
+        public PostgreSqlComparisonIsGreaterThanOrEqualTests()
+        {
+            ValidateUInt128 = false;
+        }
+
+        public override IQueryable<TestClass> GetTestList()
+        {
+            return PostgreSqlHelper.GetTestList();
+        }
+    }
+}