+++ src/tests/testfiles/PostgreSqlComparisonIsLessThanOrEqualTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceQuery.Xunit
+{
+    public class PostgreSqlComparisonIsLessThanOrEqualTests : ComparisonIsLessThanOrEqualTests<TestClass>
+    {
+        public PostgreSqlComparisonIsLessThanOrEqualTests()
+        {
+            ValidateUInt128 = false;
+        }
+
+        public override IQueryable<TestClass> GetTestList()
+        {
+            return PostgreSqlHelper.GetTestList();
+        }
+    }
+}