+++ src/tests/testfiles/PostgreSqlDistinctTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceQuery.Xunit
+{
+    public class PostgreSqlDistinctTests : DistinctTests<TestClass>
+    {
+        public PostgreSqlDistinctTests()
+        {
+            ValidateUInt128 = false;
+        }
+
+        public override IQueryable<TestClass> GetTestList()
+        {
+            return PostgreSqlHelper.GetTestList();
+        }
+    }
+}