+++ ReleaseNotes.md
@@ -1,2 +1,6 @@
-# 1.0.0-alpha (2022-08-13)
-- Alpha testing release
+# 1.0.6
+- DateTime parsing using System.Globalization.DateTimeStyles.RoundtripKind
+
+# 1.0.5
+- Full source code release with unit and integration tests
+